Remove deprecated model generator
